feat: implement Actor and ActorAction classes with specific actions for Ambulancia, Operario, and Asociado
