+++ src/mdp/ingenieria/clinicagestion/model/simulation/actor/AsociadoActor.java
@@ -1,24 +1,42 @@
-package mdp.ingenieria.clinicagestion.model.actor;
+package mdp.ingenieria.clinicagestion.model.simulation.actor;
 
 import java.util.Queue;
 
-import mdp.ingenieria.clinicagestion.model.actor.actoraction.AsociadoActorAction;
 import mdp.ingenieria.clinicagestion.model.persona.Asociado;
+import mdp.ingenieria.clinicagestion.model.simulation.Actor;
+import mdp.ingenieria.clinicagestion.model.simulation.actoraction.AsociadoActorAction;
 
-public class AsociadoActor extends Thread {
+public class AsociadoActor extends Actor {
 
 	private Asociado asociado;
 	
 	private Queue<AsociadoActorAction> actionQueue;
 	
 	public AsociadoActor(Asociado asociado, Queue<AsociadoActorAction> actionQueue) {
+		super();
 		assert asociado != null;
 		assert actionQueue != null;
 		this.asociado = asociado;
 		this.actionQueue = actionQueue;
 	}
-
 	
+	@Override
+	public void run() {
+		while (!actionQueue.isEmpty() && this.isSimulationRunning())
+		{	
+			AsociadoActorAction action = this.actionQueue.poll();
+			action.execute( this.asociado );
+		}
+	}
+	
+	public void addActorAction (AsociadoActorAction action) {
+		this.actionQueue.add(action);
+	}
+	
+	public void clearActions () {
+		this.actionQueue.clear();
+	}
+
 	/**
 	 * cuando el usuario interactua con la simulacion, se deberia crear otro AsociadoActor con
 	 * la implementacion actual, ya que los hilos viven hasta que se completen sus tareas.
@@ -36,21 +54,4 @@ public class AsociadoActor extends Thread {
 	 * Retorno automático a la clínica (evento temporal del sistema).
 	 * creo que esto podria solucionarlo 
 	 */
-	
-	@Override
-	public void run() {
-		while (!actionQueue.isEmpty())
-		{	
-			AsociadoActorAction action = this.actionQueue.poll();
-			action.execute( this.asociado );
-		}
-	}
-	
-	public void addActorAction (AsociadoActorAction action) {
-		this.actionQueue.add(action);
-	}
-	
-	public void clearActions () {
-		this.actionQueue.clear();
-	}
 }
